Db ids to int

+++ app/src/main/java/ch/band/inf2019/uk335/db/Categorie.java
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey;
 
 @Entity(tableName = "categorie_table")
 public class Categorie {
-    @PrimaryKey(autoGenerate = true) public long id;
+    @PrimaryKey(autoGenerate = true) public int id;
 
     public String title;
